+++ src/markets/NZDUSD/params.go
@@ -2,12 +2,12 @@ package NZDUSD
 
 import "TradingBot/src/types"
 
-var priceAdjustment float32 = float32(1) / float32(10000)
+var priceAdjustment float64 = float64(1) / float64(10000)
 
 var EMACrossoverLongParams = types.MarketStrategyParams{
 	RiskPercentage: 3,
 
-	StopLossPriceOffset: float64(0 * priceAdjustment),
+	StopLossPriceOffset: 0 * priceAdjustment,
 	CandlesAmountForHorizontalLevel: &types.CandlesAmountForHorizontalLevel{
 		Future: 20,
 		Past:   0,
@@ -17,12 +17,12 @@ var EMACrossoverLongParams = types.MarketStrategyParams{
 	TakeProfitDistance:  200 * priceAdjustment,
 	MinProfit:           99999 * priceAdjustment,
 	TrailingStopLoss: &types.TrailingStopLoss{
-		TPDistanceShortForTighterSL: float64(15 * priceAdjustment),
-		SLDistanceWhenTPIsVeryClose: float64(15 * priceAdjustment),
+		TPDistanceShortForTighterSL: 15 * priceAdjustment,
+		SLDistanceWhenTPIsVeryClose: 15 * priceAdjustment,
 	},
 	TrailingTakeProfit: &types.TrailingTakeProfit{
-		SLDistanceShortForTighterTP: float64(140 * priceAdjustment),
-		TPDistanceWhenSLIsVeryClose: float64(-100 * priceAdjustment),
+		SLDistanceShortForTighterTP: 140 * priceAdjustment,
+		TPDistanceWhenSLIsVeryClose: -100 * priceAdjustment,
 	},
 	CandlesAmountWithoutEMAsCrossing: 2,
 	MaxSecondsOpenTrade:              0,
@@ -34,7 +34,7 @@ var EMACrossoverLongParams = types.MarketStrategyParams{
 var EMACrossoverShortParams = types.MarketStrategyParams{
 	RiskPercentage: 3,
 
-	StopLossPriceOffset: float64(60 * priceAdjustment),
+	StopLossPriceOffset: 60 * priceAdjustment,
 	CandlesAmountForHorizontalLevel: &types.CandlesAmountForHorizontalLevel{
 		Future: 15,
 		Past:   28,
@@ -44,8 +44,8 @@ var EMACrossoverShortParams = types.MarketStrategyParams{
 	TakeProfitDistance:  500 * priceAdjustment,
 	MinProfit:           370 * priceAdjustment,
 	TrailingStopLoss: &types.TrailingStopLoss{
-		TPDistanceShortForTighterSL: float64(100 * priceAdjustment),
-		SLDistanceWhenTPIsVeryClose: float64(40 * priceAdjustment),
+		TPDistanceShortForTighterSL: 100 * priceAdjustment,
+		SLDistanceWhenTPIsVeryClose: 40 * priceAdjustment,
 	},
 	CandlesAmountWithoutEMAsCrossing: 27,
 	MaxSecondsOpenTrade:              0,
